requirements.txt
setuptools>=65.0.0
wheel
Flask==2.2.2
Pillow>=12.0.0
numpy>=1.26.0,<2.0
scikit-learn>=1.2.2
Werkzeug==2.3.7